Fix evaluate stage for custom transformer
This updates the DVC pipelines which now is
complete.
