+++ Dvcfile
@@ -1,20 +1,20 @@
-md5: 220d0157b4c69cd45c029c3ac99333ad
+md5: f500588828afffe43395829f24392bc3
 cmd: python src/evaluate.py data/prepared-data-test.csv models/misog-model.joblib
   reports
 deps:
-- md5: 981da9e4e1849a5834cb96070a6c3376
+- md5: 14552595b62af11d8f4addb73c73719c
   path: src/evaluate.py
 - md5: 1ffd2f30f59d3c9772d7aa631dcb2e6f
   path: data/prepared-data-test.csv
-- md5: 758336a6be2a6b02374a85f037c967af
+- md5: fbd8089edade2ec592ede83bc3796326
   path: models/misog-model.joblib
 outs:
-- md5: bd47f058c292407212377d955c9a27b0
+- md5: 67422b89c1f77ad9db007ff8c154b25e
   path: reports/roc_auc_f1.png
   cache: true
   metric: false
   persist: false
-- md5: 9d84b46ee206ef4fa2365169fc20ee0e
+- md5: da46548b19cd3fbd43673179b416da47
   path: reports/eval.json
   cache: false
   metric: true
