+++ reports/eval.json
@@ -1,4 +1,4 @@
 {
-    "f1": 0.6663944467129441,
-    "AUC": 0.8280540493899066
+    "f1": 0.7218216318785577,
+    "AUC": 0.8518335544430732
 }
