+++ src/train.py
@@ -1,32 +1,15 @@
-# pylint: disable=C0103,W0613,W0201
+# pylint: disable=C0103,W0613,W0201,W0611
 import logging
 import sys
-import joblib
-import numpy as np
+import cloudpickle
 import pandas as pd
-import spacy
-from sklearn.base import BaseEstimator, TransformerMixin
-from sklearn.utils.validation import check_is_fitted
+from sklearn.experimental import enable_hist_gradient_boosting  # noqa
 from sklearn.pipeline import make_pipeline
-from sklearn.ensemble import RandomForestClassifier
+from sklearn.ensemble import HistGradientBoostingClassifier
 
+from src.transformers import SpacyTransformer
 
 logger = logging.getLogger(__name__)
-
-
-class SpacyTransformer(BaseEstimator, TransformerMixin):
-    def __init__(self):
-        pass
-
-    def fit(self, X, y):
-        self.nlp_ = spacy.load("en_core_web_md")
-        return self
-
-    def transform(self, X, y=None):
-        check_is_fitted(self)
-        docs = list(self.nlp_.pipe(X))
-        feature_matrix = np.array(list(map(lambda x: x.vector, docs)))
-        return feature_matrix
 
 
 def main():
@@ -47,8 +30,9 @@ def main():
     # docs = list(nlp.pipe(df_in["text"]))
     # feature_matrix = np.array(list(map(lambda x: x.vector, docs)))
 
-    classifier = RandomForestClassifier(
-        n_estimators=400, random_state=42, n_jobs=-1, verbose=2
+    classifier = HistGradientBoostingClassifier(
+        verbose=2,
+        # n_estimators=400, random_state=42, n_jobs=-1, verbose=2
     )
     pipeline = make_pipeline(SpacyTransformer(), classifier)
     pipeline.fit(df_in["text"], df_in["label"])
@@ -56,7 +40,7 @@ def main():
     # print(f"Input: {input_file}")
     # print(f"Output: {output_file}")
     # with open(input_file, "rb") as handler:
-    #     feature_matrix = joblib.load(handler)
+    #     feature_matrix = cloudpickle.load(handler)
     # labels = feature_matrix[:, -1]
     # feature_matrix = feature_matrix[:, :-1]
     # print(f"The feature matrix has dimensions {feature_matrix.shape}.")
@@ -67,7 +51,7 @@ def main():
     # print("Training model...")
     # classifier.fit(x_train, labels)
     with open(output_file, "wb") as handler:
-        joblib.dump(pipeline, handler, compress="zlib")
+        cloudpickle.dump(pipeline, handler)  # , compress="zlib")
 
 
 if __name__ == "__main__":
